% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add2usedObj.R
\docType{methods}
\name{add2usedObj}
\alias{add2usedObj}
\alias{add2usedObj,add2usedObj-method}
\title{description of function add2usedObj}
\usage{
add2usedObj(dataObj, list.name, data.name, object)
}
\arguments{
\item{dataObj}{the StefansExpressionSet object}

\item{list.name}{the toplevel name in the usedObj slot to store the data in}

\item{data.name}{the second level data name (the first level becomes a list; default = NULL)}

\item{object}{the object to store in the usedObj}
}
\description{
Add values to the usedObj list including some checks
}

