% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StefansExpressionSet.R
\docType{methods}
\name{addAnnotation}
\alias{addAnnotation}
\alias{addAnnotation,StefansExpressionSet-method}
\title{description of function addAnnotation}
\usage{
addAnnotation(x, mart, mart.col = "refseq_mrna")
}
\arguments{
\item{x}{the StefansExpressionSet object}

\item{mart}{the annotation table (data.frame or mart object)}

\item{mart.col}{which column corresponds to the rownames(x@data)}
}
\description{
add aditional annotation to the StefansExpressionSet
}

