% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpressionSet.R
\name{addAnnotation.NGSexpressionSet}
\alias{addAnnotation.NGSexpressionSet}
\title{add aditional annotation to the ExpressionSet}
\usage{
\method{addAnnotation}{NGSexpressionSet}(x, mart, mart.col = "refseq_mrna")
}
\arguments{
\item{x}{the ExpressionSet object}

\item{mart}{the annotation table (data.frame or mart object)}

\item{mart.col}{which column corresponds to the rownames(x$data)}
}
\description{
add aditional annotation to the ExpressionSet
}

