% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.gg.melt.addGeneColGroup.R
\docType{methods}
\name{addGeneColGroup}
\alias{addGeneColGroup}
\alias{addGeneColGroup,StefansExpressionSet-method}
\title{description of function addGeneColGroup}
\usage{
addGeneColGroup(x, melted, colName = NULL)
}
\arguments{
\item{x}{the StefansExpressionSet}

\item{melted}{the melted table}

\item{colName}{the annotation column names to use for the coloring default=NULL}
}
\description{
an internal function to add row level clusters to a merged data table
}

