% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StefansExpressionSet.R
\docType{methods}
\name{addSampleColGroup}
\alias{addSampleColGroup}
\alias{addSampleColGroup,StefansExpressionSet-method}
\title{description of function addSampleColGroup}
\usage{
addSampleColGroup(x, melted, colName = NULL)
}
\arguments{
\item{x}{the StefansExpressionSet object}

\item{melted}{the already melted matrix}

\item{colName}{the colnames to add to the matrix (defult=NULL)}
}
\description{
an internal method adding new color columns to the melted dataset
}

