% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anayse_working_set.R
\docType{methods}
\name{anayse_working_set}
\alias{anayse_working_set}
\alias{anayse_working_set,NGSexpressionSet-method}
\title{description of function anayse_working_set}
\usage{
anayse_working_set(dataOBJ, name, p_values = c(0.1, 0.01, 0.001, 1e-04, 1e-05,
  1e-06, 1e-07, 1e-08), geneNameCol = "mgi_symbol", batchCorrect = NULL)
}
\arguments{
\item{m}{the NGSexpressionSet}
}
\value{
The z scored NGSexpressionSet
}
\description{
zscore the ngs dataset
}
