% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.NGSexpressionSet.R
\docType{methods}
\name{as_NGSexpressionSet}
\alias{as_NGSexpressionSet}
\title{description of function as_NGSexpressionSet}
\usage{
as_NGSexpressionSet(dat)
}
\arguments{
\item{dat}{the counts list you get from Rsubread::featureCounts()}
}
\description{
create a NGSexpressionSet from a counts list object
}
