% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_check.depth.R
\docType{methods}
\name{check.depth}
\alias{check.depth}
\alias{check.depth,NGSexpressionSet-method}
\title{description of function check}
\usage{
check.depth(x, genes = NULL, cutoff = 0.77)
}
\arguments{
\item{x}{the NGSexpressionSet}

\item{cutoff}{a cutoff that defines how many reads may be taken from the top 5 percent genes before a sample gets rejected}

\item{gene}{an optional list of genes you want to select for}
}
\value{
a vector of sample names that did not pass the check
}
\description{
this function checks how many reads are taken from the total dataset using the top 5 percent of genes only
}
