% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGSexpressionSet.R
\name{check.NGSexpressionSet}
\alias{check.NGSexpressionSet}
\title{this check is testing how many percent of the total reads are taken from the first x percent of the genes}
\usage{
\method{check}{NGSexpressionSet}(x, genes = NULL, cutoff = 0.77)
}
\arguments{
\item{x}{the NGSexpressionSet}

\item{genes}{see reads.taken}

\item{cutoff}{= 0.77 a good expression dataset should not use more than 77% of the reads in the top 5%

this function calls reads.taken.NGSexpressionSet internally.}
}
\value{
a list of samples which have passed the test
}
\description{
this check is testing how many percent of the total reads are taken from the first x percent of the genes
}

