% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpressionSet_Cluster.R
\name{clusters}
\alias{clusters}
\title{cclusters calculates all possible clusters on the dataset
supported are hclust cclust and tclust with there respective options}
\usage{
clusters(dataObj, clusterby = "raw", mds.proj = NULL, groups.n = 3,
  ctype = "hierarchical clust", onwhat = "Expression", cmethod = "ward.D",
  tc_restr = "eigen", tc_alpha = 0.05, tc_nstart = 50, tc_iter.max = 20,
  tc_restr.fact = 20)
}
\arguments{
\item{dataObj}{the ExpressionSet object}

\item{clusterby}{(raw or MDS)}
}
\description{
cclusters calculates all possible clusters on the dataset
supported are hclust cclust and tclust with there respective options
}

