% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.coexprees4gene.R
\docType{methods}
\name{coexprees4gene}
\alias{coexprees4gene}
\alias{coexprees4gene,StefansExpressionSet-method}
\title{description of function coexprees4gene}
\usage{
coexprees4gene(x, gene = NULL, method = "spearman",
  geneNameCol = "gene_name", padjMethod = "BH")
}
\arguments{
\item{x}{the StefansExpressionSet varibale}

\item{method}{any method supported by \link[=cor.test]{cor.test}}

\item{geneNameCol}{the name of the gene column (gene_name)}

\item{padjMethod}{the method to calucate the FDR with \link[=p.adjust]{p.adjust}}
}
\value{
a data.frame with the columns 'GeneSymbol', 'pval', 'cor', 'adj.p'
}
\description{
Calculate the coexpression for any given gene
}

