% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpressionSet.R
\name{coexprees4gene}
\alias{coexprees4gene}
\title{Calculate the coexpression for any given gene}
\usage{
coexprees4gene(x, gene = NULL, method = "spearman",
  geneNameCol = "gene_name", padjMethod = "BH")
}
\arguments{
\item{x}{the ExpressionSet varibale}

\item{method}{any method supported by \link[=cor.test]{cor.test}}

\item{geneNameCol}{the name of the gene column (gene_name)}

\item{padjMethod}{the method to calucate the FDR with \link[=p.adjust]{p.adjust}}
}
\value{
a data.frame with the columns 'GeneSymbol', 'pval', 'cor', 'adj.p'
}
\description{
Calculate the coexpression for any given gene
}

