% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StefansExpressionSet.R
\docType{methods}
\name{collaps}
\alias{collaps}
\alias{collaps,StefansExpressionSet-method}
\title{description of function collaps}
\usage{
collaps(dataObj, by = c("median", "mean", "sd", "sum"),
  groupCol = "GroupID")
}
\arguments{
\item{dataObj}{the StefansExpressionSet}

\item{by}{collapsing method c('median','mean','sd','sum', or own function )}

\item{groupCol}{the sample names you want to group on}
}
\description{
This function will collpase the data in the StefansExpressionSet to only contain one value

per sample group.
}

