% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGSexpressionSet.R
\name{collaps.NGSexpressionSet}
\alias{collaps.NGSexpressionSet}
\title{merge the groups in the dataset into one single column}
\usage{
\method{collaps}{NGSexpressionSet}(dataObj, what = c("median", "mean", "sd",
  "sum"))
}
\arguments{
\item{dataObj}{the NGSexpressionSet}

\item{what}{merge by 'median','mean','sd' or 'sum'

The sample description is restricted to the first entry in each group.
Most variables in the sample description table might be useless after this collape}
}
\value{
the collapsed NGSexpressionSet
}
\description{
merge the groups in the dataset into one single column
}

