% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.complexHeatmap.R
\docType{methods}
\name{complexHeatmap}
\alias{complexHeatmap}
\alias{complexHeatmap,StefansExpressionSet-method}
\title{description of function plot.beans}
\usage{
complexHeatmap(x, ofile = NULL, colGroups = NULL, rowGroups = NULL,
  colColors = NULL, rowColors = NULL, pdf = FALSE, subpath = "",
  main = "", heapmapCols = function(x) {     c("darkgrey", bluered(x)) },
  brks = 10)
}
\arguments{
\item{x}{the StefansExpressionSet object}

\item{ofile}{the outfile to create in the x@outpath folder}

\item{colGroups}{columns in the samples table to use to order the data (first == order)}

\item{rowGroups}{rows in the annotation table to use to color the heatmap rows (first == order)}

\item{colColors}{a named list of column color vectors}

\item{rowColors}{a named list of row color vectors}

\item{pdf}{export as pdf (default = FALSE)}

\item{subpath}{the subpath for the plots (default = '')}

\item{heapmapCols}{the color function to calculate the heatmap colours ( default function (x) { c("darkgrey",bluered(x)) } )}

\item{brks}{how many breaks should the expression value color key have (default=10)}
}
\description{
plot the PCR heatmap using the heatmap.3 function included in this package
}
