% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.cor2cytoscape.R
\docType{methods}
\name{cor2cytoscape}
\alias{cor2cytoscape}
\alias{cor2cytoscape,StefansExpressionSet-method}
\title{description of function cor2cytoscape}
\usage{
cor2cytoscape(M, file, cut = 0.9)
}
\arguments{
\item{M}{the correlation matrix obtained by a run of \code{\link{corMat}}}
}
\description{
exports the correlation matrix in the format \url{http://www.cytoscape.org/} does

support as network file.
}
