% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpressionSet.R
\name{cor2cytoscape}
\alias{cor2cytoscape}
\title{exports the correlation matrix in the format \href{http://www.cytoscape.org/}{CytoScape} 
does support as network file.}
\usage{
cor2cytoscape(M, file, cut = 0.9)
}
\arguments{
\item{M}{the correlation matrix obtained by a run of \code{\link{corMat}}}
}
\description{
exports the correlation matrix in the format \href{http://www.cytoscape.org/}{CytoScape} 
does support as network file.
}

