% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpressionSet.R
\name{corMat}
\alias{corMat}
\title{calculate a correlation matrix for the data object}
\usage{
corMat(x, sd_cut = 1, method = "spearman", geneNameCol = "gene_name",
  groupCol = NULL, name = "tmp")
}
\arguments{
\item{x}{the ExpressionSet varibale}

\item{sd_cut}{the cut off value for the sd check}

\item{method}{any method supported by \link[=cor.test]{cor.test}}

\item{geneNameCol}{the name of the gene column (gene_name)}

\item{groupCol}{the column name of the grouping variable in the samples table}

\item{name}{the name of the analysis}
}
\value{
the correlation matrix
}
\description{
calculate a correlation matrix for the data object
}

