% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StefansExpressionSet.R
\docType{methods}
\name{createStats}
\alias{createStats}
\alias{createStats,StefansExpressionSet-method}
\title{description of function createStats}
\usage{
createStats(x, condition, files = F, A = NULL, B = NULL)
}
\arguments{
\item{x}{the StefansExpressionSet object}

\item{condition}{the samples column containing the condition of interest}

\item{files}{whether or not (default) export the stat files using \code{\link{writeStatTables}}}

\item{A}{an optional condition A to compare to a condition B}

\item{B}{the condition B}
}
\description{
constructor that has to be implemented in the data specific classes
}

