% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGSexpressionSet.R
\name{createStats.NGSexpressionSet}
\alias{createStats.NGSexpressionSet}
\title{calculate staistics on all possible groupings using the DEseq nbinomTest test}
\usage{
\method{createStats}{NGSexpressionSet}(x, condition, files = F, A = NULL,
  B = NULL)
}
\arguments{
\item{x}{the NGSexpressionSet}

\item{condition}{the grouping column in the samples data.frame}

\item{files}{write the statistics tables (FALSE)}

\item{A}{the first component to analyze (Group A)}

\item{B}{the second component to analyze (Group B)
Both together create the group 'A vs. B'}
}
\value{
the NGSexpressionSet with a set of ststs tables
}
\description{
calculate staistics on all possible groupings using the DEseq nbinomTest test
}

