% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpressionSet.R
\name{createStats}
\alias{createStats}
\title{constructors that have to be implemented in the classes}
\usage{
createStats(x, condition, files = F, A = NULL, B = NULL)
}
\description{
constructors that have to be implemented in the classes
}

