% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpressionSet.R
\name{createWorkingSet}
\alias{createWorkingSet}
\title{this file contains all generic fnction for data export and ploting
Create an ExpressionSet object (S3)
This object is mainly used for subsetting of the data and plotting}
\usage{
createWorkingSet(dat, Samples, Analysis = NULL, name = "WorkingSet",
  namecol = "GroupName", namerow = "GeneID", usecol = "Use",
  outpath = NULL)
}
\arguments{
\item{dat}{data frame or matrix containing all expression data}

\item{Samples}{A sample description table}

\item{Analysis}{If the samples table contains a Analysis column you can subset the data already here to the Analysis here (String). This table has to contain a column filenames that is expected to connect the sample table to the dat column names.}

\item{name}{The name of this object is going to be used in the output of all plots and tables - make it specific}

\item{namecol}{The samples table column that contains the (to be set) names of the samples in the data matrix}

\item{namerow}{This is the name of the gene level annotation column in the dat file to use as ID}

\item{outpath}{Where to store the output from the analysis}

\item{annotation}{The annotation table from e.g. affymetrix csv data}

\item{newOrder}{The samples column name for the new order (default 'Order')}
}
\description{
this file contains all generic fnction for data export and ploting
Create an ExpressionSet object (S3)
This object is mainly used for subsetting of the data and plotting
}

