% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.gg.defineHeatmapColors.R
\docType{methods}
\name{defineHeatmapColors}
\alias{defineHeatmapColors}
\alias{defineHeatmapColors,SingleCellsNGS-method}
\title{description of function gg.heatmap.list}
\usage{
defineHeatmapColors(x, melted, ..., colrs = NULL)
}
\arguments{
\item{colrs}{and optional colors vector( gray + bluered for data and rainbow for samples)}

\item{merged}{the merged data object with the Expression column that should be colored}
}
\value{
a list with the modified merged table and the colors vector
}
\description{
uses ggplot2 to plot heatmaps
}

