% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.probeset.R
\docType{methods}
\name{describe.probeset}
\alias{describe.probeset}
\alias{describe.probeset,StefansExpressionSet-method}
\title{description of function describe.probeset}
\usage{
describe.probeset(x, probeset)
}
\arguments{
\item{x}{the StefansExpressionSet object}

\item{probeset}{the probeset to describe}
}
\description{
print all annotation values + all values + all stat results for one probeset
}

