% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clusters.R
\docType{methods}
\name{difference}
\alias{difference}
\alias{difference,Clusters-method}
\title{description of function difference}
\usage{
difference(x, clusters, groups.n)
}
\arguments{
\item{x}{the StefansExpressionSet object}

\item{clusters}{the grouoing strings}

\item{groups.n}{the max number of groups}
}
\description{
this code has to be revised! NOT WORKING
}

