% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StefansExpressionSet.R
\docType{methods}
\name{drop.samples}
\alias{drop.samples}
\alias{drop.samples,StefansExpressionSet-method}
\title{description of function drop.samples}
\usage{
drop.samples(x, samplenames = NULL, name = "dropped_samples")
}
\arguments{
\item{x}{the StefansExpressionSet object}

\item{samplenames}{which samples to drop (samples like colnames(x@data))}

\item{name}{the name of the new StefansExpressionSet object}
}
\description{
drops samples from the StefansExpressionSet
}

