% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpressionSet.R
\name{drop.samples.ExpressionSet}
\alias{drop.samples.ExpressionSet}
\title{drops samples from the ExpressionSet}
\usage{
\method{drop.samples}{ExpressionSet}(x, samplenames = NULL,
  name = "dopped_samples")
}
\arguments{
\item{x}{the ExpressionSet object}

\item{samplenames}{which samples to drop (samples like colnames(x$data))}

\item{name}{the name of the new ExpressionSet object}
}
\description{
drops samples from the ExpressionSet
}

