% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StefansExpressionSet.R
\docType{methods}
\name{export.data}
\alias{export.data}
\alias{export.data,StefansExpressionSet-method}
\title{description of function write.data}
\usage{
export.data(x)
}
\arguments{
\item{x}{the StefansExpressionSet object}
}
\description{
write the StefansExpressionSet data, annotation and samples tables to disk
}

