% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpressionSet.R
\name{export4GEDI}
\alias{export4GEDI}
\title{exports the ExpressionSet in the format \href{https://apps.childrenshospital.org/clinical/research/ingber/GEDI/gedihome.htm}[GEDI] can import.}
\usage{
export4GEDI(x, fname = "GEDI_output.txt", tag.col = NULL, time.col = NULL,
  minSample_PerTime = 1)
}
\arguments{
\item{x}{the ExpressionSet object}

\item{fname}{the filename to export the data to}

\item{tag.col}{the sample name column in the samples table}

\item{time.col}{the time column in the samples table}
}
\description{
exports the ExpressionSet in the format \href{https://apps.childrenshospital.org/clinical/research/ingber/GEDI/gedihome.htm}[GEDI] can import.
}

