% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpressionSet.R
\docType{class}
\name{ExpressionSet-class}
\alias{ExpressionSet-class}
\title{An S4 class to visualize Expression data.}
\description{
An S4 class to visualize Expression data.
}
\section{Slots}{

\describe{
\item{\code{data}}{a data.frame containing the expression values for each gene x sample (gene = row)}

\item{\code{samples}}{a data.frame describing the columnanmes in the data column}

\item{\code{annotation}}{a data.frame describing the rownames in the data rows}

\item{\code{outpath}}{the default outpath for the plots and tables from this package}

\item{\code{name}}{the name for this package (all filesnames contain that)}

\item{\code{rownamescol}}{the column name in the annotation table that represents the rownames of the data table}

\item{\code{sampleNamesCol}}{the column name in the samples table that represents the colnames of the data table}

\item{\code{stats}}{the stats list for all stats created in the object}
}}

