% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StefansExpressionSet.R
\docType{methods}
\name{force.numeric}
\alias{force.numeric}
\alias{force.numeric,StefansExpressionSet-method}
\title{description of function force.numeric}
\usage{
force.numeric(dataObj)
}
\arguments{
\item{dataObj}{the StefansExpressionSet object}
}
\description{
The moethod forces the values in the data matrix to be numbers.
}

