% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpressionSet.R
\name{forceAbsoluteUniqueSample}
\alias{forceAbsoluteUniqueSample}
\title{force absolute unique names in a vector by adding _<amount of repeats> to each value if
there is more than one repeat opf the value}
\usage{
forceAbsoluteUniqueSample(x, separator = "_")
}
\arguments{
\item{x}{the ExpressionSet object}

\item{separator}{'_' or anything you want to set the separator to}
}
\description{
force absolute unique names in a vector by adding _<amount of repeats> to each value if
there is more than one repeat opf the value
}

