% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Accessors.R
\docType{methods}
\name{get_gene_list}
\alias{get_gene_list}
\alias{get_gene_list,StefansExpressionSet-method}
\title{description of function get_gene_list}
\usage{
get_gene_list(x, p_value = c(0.1, 0.01, 0.001, 1e-04, 1e-05, 1e-06, 1e-07,
  1e-08), geneNameCol = NULL)
}
\arguments{
\item{x}{the ExpressionSet}

\item{p_value}{a list of p value cut offs (0.1,  1e-2 ,1e-3,1e-4,1e-5, 1e-6, 1e-7, 1e-8 )}

\item{geneNameCol}{the (optional) annotation column defining the gene names}
}
\value{
the ExpressionSet object with values in the sig_genes slot
}
\description{
Query all stat tables and select genes based on multiple p value cutoff values
}

