% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StefansExpressionSet.R
\docType{methods}
\name{getAnnotation4probesets}
\alias{getAnnotation4probesets}
\alias{getAnnotation4probesets,StefansExpressionSet-method}
\title{description of function getAnnotation4probesets}
\usage{
getAnnotation4probesets(x, probesets = c(), colname = "Gene.Symbol")
}
\arguments{
\item{x}{the StefansExpressionSet object}

\item{probesets}{the vector of probesets to annotate (rownames(x@data))}

\item{colname}{a vector of colnames to annotate}
}
\value{
a vector of annotation values
}
\description{
subsets the annotation table for a set of probesets
}

