% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpressionSet.R
\name{getAnnotation4probesets}
\alias{getAnnotation4probesets}
\title{subsets the annotation table for a set of probesets}
\usage{
getAnnotation4probesets(x, probesets = c(), colname = "Gene.Symbol")
}
\arguments{
\item{x}{the ExpressionSet object}

\item{probesets}{the vector of probesets to annotate (rownames(x@data))}

\item{colname}{a vector of colnames to annotate}
}
\value{
a vector of annotation values
}
\description{
subsets the annotation table for a set of probesets
}

