% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGSexpressionSet.R
\name{getProbesetsFromStats.NGSexpressionSet}
\alias{getProbesetsFromStats.NGSexpressionSet}
\title{returns a list of probesets (the rownames from the data matrix) for a restriction of a list of stat comparisons}
\usage{
\method{getProbesetsFromStats}{NGSexpressionSet}(x, v = 1e-04, pos = "padj",
  mode = "less", Comparisons = NULL)

\method{getProbesetsFromStats}{NGSexpressionSet}(x, v = 1e-04, pos = "padj",
  mode = "less", Comparisons = NULL)
}
\arguments{
\item{v}{The cutoff value}

\item{pos}{The column in the stats tables to base the selection on}

\item{mode}{one of 'less', 'more', 'onlyless' or 'equals' default 'less' ( <= )}

\item{Comparisons}{A list of comparisons to check (all if left out)}

\item{v}{The cutoff value}

\item{pos}{The column in the stats tables to base the selection on}

\item{Comparisons}{A list of comparisons to check (all if left out)}

\item{mode}{one of 'less', 'more', 'onlyless' or 'equals' default 'less' ( <= )}
}
\value{
a list of probesets that shows an adjusted p value below v (1e-4 in the example)

a list of probesets that shows an adjusted p value below 1e-4
}
\description{
returns a list of probesets (the rownames from the data matrix) for a restriction of a list of stat comparisons

getProbesetsFromStats returns a list of probesets (the rownames from the data matrix) for a restriction of a list of stat comparisons
}
\examples{
probes <- getProbesetsFromStats ( x, v=1e-4, pos="adj.P.Val" ) 
probes <- getProbesetsFromStats ( x, v=1e-4, pos="adj.P.Val" ) 
}

