% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Accessors.R
\docType{methods}
\name{getProbesetsFromValues}
\alias{getProbesetsFromValues}
\alias{getProbesetsFromValues,StefansExpressionSet-method}
\title{description of function getProbesetsFromValues}
\usage{
getProbesetsFromValues(x, v = "NULL", sample = "NULL", mode = "less")
}
\arguments{
\item{x}{the ExpreesionSet object}

\item{v}{the cut off- or matching value}

\item{sample}{sample names as used in data column names}

\item{mode}{one of less, more, onlyless or equals}
}
\value{
a list of probesets that match the requirements
}
\description{
Using this function you can select probes based on there expression value in a set of samples
}

