% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGSexpressionSet.R
\name{getProbesetsFromValues.NGSexpressionSet}
\alias{getProbesetsFromValues.NGSexpressionSet}
\title{Select probesets, that show a certain level in expression for a single sample}
\usage{
\method{getProbesetsFromValues}{NGSexpressionSet}(x, v = NULL,
  sample = NULL, mode = "less")

\method{getProbesetsFromValues}{NGSexpressionSet}(x, v = NULL,
  sample = NULL, mode = "less")
}
\arguments{
\item{v}{The cutoff value}

\item{sample}{The sample name}

\item{mode}{one of 'less', 'more', 'onlyless' or 'equals' default 'less' ( <= )}

\item{v}{The cutoff value}

\item{sample}{The sample name}

\item{mode}{one of 'less', 'more', 'onlyless' or 'equals' default 'less' ( <= )}
}
\value{
a list of probesets that has a expression less than 10 in sample A

a list of probesets that has a expression less than 10 in sample A
}
\description{
Select probesets, that show a certain level in expression for a single sample

Select probesets, that show a certain level in expression for a single sample
}
\examples{
probes <- getProbesetsFromStats ( x, v=10, sample="A" ) 
probes <- getProbesetsFromStats ( x, v=10, sample="A" ) 
}

