% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ggplot.gene.R
\docType{methods}
\name{ggplot.gene}
\alias{ggplot.gene}
\alias{ggplot.gene,StefansExpressionSet-method}
\title{description of function ggplot.gene}
\usage{
ggplot.gene(dat, gene, colrs = NULL, groupCol = "GroupID", colCol = NULL,
  boxplot = F, log2 = F)
}
\arguments{
\item{dat}{the StefansExpressionSet object}

\item{gene}{the gene of interest}

\item{colrs}{the grouping colors for the x axis (samples)}

\item{groupCol}{the samples clumn that contains the grouping information}

\item{colCol}{the sample column that contains the color information}

\item{log2}{log2 transform data before plotting (default = F)}
}
\description{
Plot one gene in the StefansExpressionSet as boxplot or points plot (using ggplot2)
}
