% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grouping.hclust.R
\docType{methods}
\name{group.hclust}
\alias{group.hclust}
\alias{group.hclust,StefansExpressionSet-method}
\title{description of function group.hclust}
\usage{
group.hclust(x, groups, name, type = "genes", distfun = function(a) {    
  dist(a) }, hclustMethod = "ward.D2")
}
\arguments{
\item{x}{the StefansExpressionSet object}

\item{groups}{how many groups should I identify}

\item{name}{the new column name created as  paste( name, groups, 'groups' )}

\item{type}{a 'gene' or 'sample' grouping?}
}
\description{
Create a groping based on a hclust run like hclust(dist(x@data),method="ward.D2")
}

