\name{HSC_MPP1}
\title{Stats result table HSC_MPP1}
\description{As the statistics is calculated using the DESeq package I do not want to test this here. I rather include the stats table as data set.}
\usage{PMID25158935$stats <- list( 'HSC vs. MPP1' = HSC_MPP1 )}
\format{data.frame containing  24062 rows and 8 columns}