% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interesting_genes.R
\docType{methods}
\name{interesting_genes}
\alias{interesting_genes}
\alias{interesting_genes,SingleCellsNGS-method}
\title{description of function interesting_genes}
\usage{
interesting_genes(x, Min = 0.4, Max = 0.8, genes = 100)
}
\arguments{
\item{x}{the SingleCellsNGS object}

\item{Min}{the minimal fraction of cells expresing the gene default=0.4}

\item{Max}{the maximal fraction of cells expresing the gene default=0.8}

\item{genes}{how many genes you want to get back default=100}
}
\description{
This method uses the values created by update.measurements() and selects the most interesting subset of genes.
}

