% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log2.R
\docType{methods}
\name{log2}
\alias{log2}
\alias{log2,StefansExpressionSet-method}
\title{description of function log2}
\usage{
\S4method{log2}{StefansExpressionSet}(x)
}
\arguments{
\item{x}{the StefansExpressionSet}
}
\description{
Converts the x@data values to log2(x@data+1) values if not already done
}
