% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDS.mds.R
\docType{methods}
\name{mds}
\alias{mds}
\alias{mds,StefansExpressionSet-method}
\title{description of function mds.and.clus}
\usage{
mds(dataObj, ..., mds.type = "PCA", onwhat = "Expression")
}
\arguments{
\item{dataObj}{the StefansExpressionSet object}

\item{mds.type}{Which MDS function should be called default="PCA"}

\item{onwhat}{condense which dataset at the moment only Expression is supported default='expression'}
}
\description{
Calculates the MDS for a given MDS type and stores the 3 dimensions in the object for later use.
}

