% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpressionSet_Cluster.R
\name{mds.and.clus}
\alias{mds.and.clus}
\title{Author: Stefan Lang
this implements clustering of genes for the ExpressionSet class
the clustering options are taken from the SCExV R lib}
\usage{
mds.and.clus(dataObj, clusterby = "raw", mds.type = "PCA", groups.n,
  LLEK = 2, cmethod = "ward.D", ctype = "hierarchical clust",
  onwhat = "Expression", ...)
}
\description{
Author: Stefan Lang
this implements clustering of genes for the ExpressionSet class
the clustering options are taken from the SCExV R lib
}

