% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpressionSet.R
\docType{methods}
\name{melt,ExpressionSet-method}
\alias{melt,ExpressionSet-method}
\title{melts the object using  \code{\link[reshape2]{melt}}}
\usage{
\S4method{melt}{ExpressionSet}(data, groupcol = "GroupName",
  colCol = "GroupName", probeNames = "Gene.Symbol", na.rm = FALSE,
  value.name = "value")
}
\arguments{
\item{groupcol}{the grouping column in the samples data}

\item{colCol}{the coloring column in the sample data}

\item{probeNames}{the column in the annotation datacontaining the gene symbol}

\item{x}{the ExpressionSet object}
}
\description{
melts the object using  \code{\link[reshape2]{melt}}
}

