% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StefansExpressionSet.R
\docType{methods}
\name{melt}
\alias{melt}
\alias{melt,StefansExpressionSet-method}
\title{description of function melt}
\usage{
melt.StefansExpressionSet(dat, groupcol = "GroupName", colCol = "GroupName",
  probeNames = NULL, na.rm = FALSE, value.name = "value")
}
\arguments{
\item{groupcol}{the column in the samples table to group the expression on}

\item{colCol}{the column in the samples table to color the grouping data on}

\item{probeNames}{which probenames to use (ProbeSetID or Gene.Symbol ...)}

\item{data}{the StefansExpressionSet object}
}
\description{
met an StefansExpressionSet to be plotted using ggplot2 functions
}

