% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpressionSet.R
\name{melt.ExpressionSet}
\alias{melt.ExpressionSet}
\title{melts the object using  \code{\link[reshape2]{melt}}}
\usage{
melt.ExpressionSet(x, groupcol = "GroupName", colCol = "GroupName",
  probeNames = "Gene.Symbol")
}
\arguments{
\item{x}{the ExpressionSet object}

\item{groupcol}{the grouping column in the samples data}

\item{colCol}{the coloring column in the sample data}

\item{probeNames}{the column in the annotation datacontaining the gene symbol}
}
\description{
melts the object using  \code{\link[reshape2]{melt}}
}

