% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Accessors.R
\docType{methods}
\name{name_4_IDs}
\alias{name_4_IDs}
\alias{name_4_IDs,StefansExpressionSet-method}
\title{description of function name_4_IDs}
\usage{
name_4_IDs(x, ids = NULL, geneNameCol = "mgi_symbol")
}
\arguments{
\item{v}{The cutoff value}

\item{sample}{The sample name}

\item{mode}{one of 'less', 'more', 'onlyless' or 'equals' default 'less' ( <= )}
}
\value{
a list of probesets that has a expression less than 10 in sample A
}
\description{
Select probesets, that show a certain level in expression for a single sample probes
}

