% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StefansExpressionSet.R
\docType{methods}
\name{normalize}
\alias{normalize}
\alias{normalize,StefansExpressionSet-method}
\title{description of function normalize}
\usage{
normalize(x)
}
\arguments{
\item{x}{the StefansExpressionSet object}
}
\description{
constructor that has to be implemented in the data specific classes
}

