% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGSexpressionSet.R
\name{normalize.NGSexpressionSet}
\alias{normalize.NGSexpressionSet}
\title{normalize the expression data}
\usage{
\method{normalize}{NGSexpressionSet}(x, readCounts = NULL,
  to_gene_length = FALSE, geneLengthCol = "transcriptLength")
}
\arguments{
\item{x}{The NGSexpressionSet}

\item{readCounts}{The number of reads from each bam file or another value you want to normalize the data to}

\item{to_gene_length}{FALSE whether or not to normalize the data to gene length}

\item{geneLengthCol}{the column in the annotation data.frame to normalize the genes to}
}
\value{
the normalized data set (original data stored in NGS$raw
}
\description{
normalize the expression data
}

