% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.pcorMat.R
\docType{methods}
\name{pcorMat}
\alias{pcorMat}
\alias{pcorMat,StefansExpressionSet-method}
\title{description of function corMat.Pvalues}
\usage{
pcorMat(x, sd_cut = 1, method = "spearman", geneNameCol = "gene_name",
  groupCol = NULL, name = "tmp")
}
\arguments{
\item{x}{the StefansExpressionSet varibale}

\item{sd_cut}{the cut off value for the sd check}

\item{method}{any method supported by \code{\link[stats]{cor.test}}}

\item{geneNameCol}{the name of the gene column (gene_name)}

\item{groupCol}{the column name of the grouping variable in the samples table}

\item{name}{the name of the analysis}
}
\description{
calculate a p_value matrix for the data object THIS FUNCTION IS NOT DOING THE RIGTH

THING BROKEN
}
