% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.stats.as.heatmap.R
\docType{methods}
\name{plot.stats.as.heatmap}
\alias{plot.stats.as.heatmap}
\alias{plot.stats.as.heatmap,StefansExpressionSet-method}
\title{description of function plot}
\usage{
plot.stats.as.heatmap(x, pvalue = c(0.1, 0.01, 0.001, 1e-04, 1e-05, 1e-06,
  1e-07, 1e-08), Subset = NULL, Subset.name = NULL, comp = NULL,
  gene_centered = F, collaps = NULL, geneNameCol = "mgi_symbol")
}
\arguments{
\item{x}{the StefansExpressionSet}

\item{pvalue}{a vector of pvalues to set as cut off}

\item{Subset}{no idear at the moment....}

\item{Subset.name}{no idear what that should do here....}

\item{comp}{a list of comparisons to restric the plotting to (NULL = all)}

\item{gene_centered}{collapse all genes with the same gene symbol into one value}

\item{collaps}{how to collapse the data if gene_centered}

\item{geneNameCol}{the name of the gene.symbol column in the annotation}
}
\description{
create heatmap for each statistics table and each pvalue cutoff This function uses

internally the plot.heatmaps() function for each selected probeset list
}
