% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.heatmaps.R
\docType{methods}
\name{plot.heatmaps}
\alias{plot.heatmaps}
\alias{plot.heatmaps,StefansExpressionSet-method}
\title{description of function plot.heatmaps}
\usage{
plot.heatmaps(dataOBJ, groupCol = "GroupName", gene.names = NULL,
  pvalue = 1, analysis_name = NULL, Subset = NULL, collaps = NULL,
  pdf = F, ...)
}
\arguments{
\item{dataOBJ}{the StefansExpressionSet object}

\item{groupCol}{the column in the samples table to group the samples on}

\item{gene.names}{the rownames(dataObj@data) level gene names}

\item{pvalue}{an optional cut off value to select genes from the statistical results tables}

\item{analysis_name}{the name for the outfiles}

\item{Subset}{an optional list of strings matching to the gene symbols used to select genes of interest}

\item{collaps}{collaps the sample groups into a single column of the heatmap using one of collaps=c('median', 'mean')}
}
\description{
An rather ancient, but flexible heatmapping function that depends on heatmap.2 and allows for many data

selection/conversion options
}

