% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.histogram.R
\docType{methods}
\name{plot.histogram}
\alias{plot.histogram}
\alias{plot.histogram,StefansExpressionSet-method}
\title{description of function plot.histograms}
\usage{
plot.histogram(dataObj, probesetID, cuts = vector("list", 1),
  subpath = "preprocess", colGroup = "ArrayID", nameCol = "gene_name",
  png = FALSE, breaks = 15)
}
\arguments{
\item{dataObj}{the StefansExpressionSet object}

\item{probesetID}{the probeset id of the gene to plot}

\item{cuts}{the cuts are used for the 1D gene groups default=vector('list',1)}

\item{subpath}{the subpath to plot to ( default = preprocess)}

\item{colGroup}{the samples table column to color the plot}

\item{nameCol}{the gene name column to enhance the plot information}

\item{png}{create a png file (default =F)}

\item{breaks}{the amount of breaks in the hist default=15}
}
\description{
This function plots one gene as histogram to check whether there ar4e clear expression differences in different plates.
}

