% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGSexpressionSet.R
\name{plot.NGSexpressionSet}
\alias{plot.NGSexpressionSet}
\title{create heatmap for each statistics table and each pvalue cutoff}
\usage{
\method{plot}{NGSexpressionSet}(x, pvalue = c(0.1, 0.01, 0.001, 1e-04, 1e-05,
  1e-06, 1e-07, 1e-08), Subset = NULL, Subset.name = NULL, comp = NULL,
  gene_centered = F, collaps = NULL, geneNameCol = "mgi_symbol")
}
\arguments{
\item{x}{the NGSexpressionSet}

\item{pvalue}{a vector of pvalues to set as cut off}

\item{Subset}{no idear at the moment....}

\item{Subset.name}{no idear what that should do here....}

\item{comp}{a list of comparisons to restric the plotting to (NULL = all)}

\item{gene_centered}{collapse all genes with the same gene symbol into one value}

\item{collaps}{how to collapse the data if gene_centered}

\item{geneNameCol}{the name of the gene.symbol column in the annotation
This function uses internally the plot.heatmaps() function for each selected probeset list}
}
\description{
create heatmap for each statistics table and each pvalue cutoff
}

