% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.probeset.R
\docType{methods}
\name{plot.probeset}
\alias{plot.probeset}
\alias{plot.probeset,StefansExpressionSet-method}
\title{description of function plot.probeset}
\usage{
plot.probeset(x, probeset, boxplot = F, pdf = F,
  geneNameCol = "mgi_symbol", sampleGroup = "GroupName")
}
\arguments{
\item{x}{the StefansExpressionSet object}

\item{probeset}{the probeset name (rownames(x@data))}

\item{boxplot}{(F or T) create a a dots- or box-plot}

\item{pdf}{save the file as pdf (default svg)}

\item{geneNameCol}{the column name for the gene symbol to use in the plots}

\item{sampleGroup}{the sample table column containing the grouping information}

\item{groupcol}{the grouping column in the samples data}

\item{colCol}{the coloring column in the sample data}

\item{probeNames}{the column in the annotation datacontaining the gene symbol}

\item{x}{the StefansExpressionSet object}
}
\description{
This function plots the expression data grouped by the GroupName using ggplot2.

Works only for one probeset.
}
\examples{
p <- plot.probeset( red, rownames(red@data)[20], geneNameCol= 'GeneID', boxplot=T)
#produces the file '../../tmp/reducedSet_boxplot_Mybl1_expression.svg'
p <- plot.probeset( red, rownames(red@data)[20], geneNameCol= 'GeneID', boxplot=F)
#produces the file '../../tmp/reducedSet_points_Mybl1_expression.svg'
}
