% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpressionSet.R
\name{plot.probeset}
\alias{plot.probeset}
\title{plot grouped probesets creates ONE plot for ONE group of probesets
If you want a multi group plot create it yourself from the single ones.}
\usage{
plot.probeset(x, probeset, boxplot = F, pdf = F,
  geneNameCol = "mgi_symbol")
}
\arguments{
\item{x}{the ExpressionSet object}

\item{probeset}{the probeset name (rownames(x@data))}

\item{boxplot}{(F or T) create a a dots- or box-plot}

\item{pdf}{save the file as pdf (default svg)}

\item{geneNameCol}{the column name for the gene symbol to use in the plots}
}
\description{
plot grouped probesets creates ONE plot for ONE group of probesets
If you want a multi group plot create it yourself from the single ones.
}

