\name{PMID25158935samples}
\title{Sample data from PMID 25158935}
\description{The sample data has been parsed from the information available on the NCBI servers}
\usage{PMID25158935samples}
\format{data.frame containing  15 samples and 21 tags}
\source{http://www.ncbi.nlm.nih.gov/bioproject/PRJEB5296}
\references{
Cabezas-Wallscheid N (2014) \emph{Identification of regulatory networks in HSCs and their immediate progeny via integrated proteome, transcriptome, and DNA methylome analysis.}
Cell Stem Cell
}

