% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.preprocess.R
\docType{methods}
\name{preprocess}
\alias{preprocess}
\alias{preprocess,NGSexpressionSet-method}
\title{description of function preprocess}
\usage{
preprocess(x, condition)
}
\arguments{
\item{x}{the NGSexpressionSet}

\item{condition}{the column in the samples table to use as grouing variable}
}
\description{
create the count dataset and normalize it following DEseq standard

function used by DEseq Do not use
}

