% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpressionSet.R
\name{print.ExpressionSet}
\alias{print.ExpressionSet}
\title{print the ExpressionSet}
\usage{
\method{print}{ExpressionSet}(x)
}
\arguments{
\item{x}{the ExpressionSet object}
}
\value{
nothing
}
\description{
print the ExpressionSet
}
\examples{

}

