% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clusters.R
\docType{methods}
\name{quality_of_fit}
\alias{quality_of_fit}
\alias{quality_of_fit,Clusters-method}
\title{description of function quality_of_fit}
\usage{
quality_of_fit(x, what = "cells", col = "groupID")
}
\arguments{
\item{x}{the StefansExpressionSet object}

\item{what}{where to group the data on (cells or genes; default='cells')}

\item{col}{which samples/annoation column to use for the grouping (default='groupID')}
}
\description{
this calculated a quality of fit for the expression data
}

