% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpressionSet.R
\name{rank.ExpressionSet}
\alias{rank.ExpressionSet}
\title{creates a new ranks vector in the ExpressionSet}
\usage{
\method{rank}{ExpressionSet}(x)
}
\arguments{
\item{x}{the ExpressionSet object}
}
\description{
creates a new ranks vector in the ExpressionSet
}

