% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_ranks.R
\docType{methods}
\name{ranks}
\alias{ranks}
\alias{ranks,StefansExpressionSet-method}
\title{description of function ranks}
\usage{
ranks(x)
}
\arguments{
\item{x}{the StefansExpressionSet object}
}
\description{
creates a new ranks vector in the StefansExpressionSet
}

