% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpressionSet.R
\name{ranks}
\alias{ranks}
\title{creates a new ranks vector in the ExpressionSet}
\usage{
ranks(x)
}
\arguments{
\item{x}{the ExpressionSet object}
}
\description{
creates a new ranks vector in the ExpressionSet
}

