% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.bams.R
\docType{methods}
\name{read.bams}
\alias{read.bams}
\alias{read.bams,NGSexpressionSet-method}
\title{description of function read.bams}
\usage{
read.bams(bamFiles, annotation, GTF.featureType = "exon",
  GTF.attrType = "gene_id", isPairedEnd = FALSE, nthreads = 2,
  as.obj = F)
}
\arguments{
\item{bamFiles}{a file containing one file name per line}

\item{annotation}{the gene level annotation to use}

\item{as.obj}{if true return the couts object instead of a matrix (default =F)}
}
\description{
Use most of'StefansExpressionSet'functionallity with minor changes to NGS data (like normalizing)

This package is mainly meant to plot the data - all important quantification or gene

annotation is performed using DEseq functionallity. Read a set of bam files and perform

the quantification (better do that without using this function!)
}
