% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGSexpressionSet.R
\name{read.bams}
\alias{read.bams}
\title{Use most of ExpressionSet functionallity with minor changes to NGS data (like normalizing)}
\usage{
read.bams(bamFiles, annotation, GTF.featureType = "exon",
  GTF.attrType = "gene_id", isPairedEnd = FALSE, nthreads = 2)
}
\arguments{
\item{bamFiles}{a file containing one file name per line}

\item{annotation}{the gene level annotation to use}
}
\description{
This package is mainly meant to plot the data - all important quantification or gene annotation is performed using 
DEseq functionallity.
}
\details{
Read a set of bam files and perform the quantification (better do that without using this function!)
}

