% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_reads.taken.R
\docType{methods}
\name{reads.taken}
\alias{reads.taken}
\alias{reads.taken,NGSexpressionSet-method}
\title{description of function reads.taken}
\usage{
reads.taken(x, percentile = 0.05, tf = NULL)
}
\arguments{
\item{x}{the NGSexpressionSet}

\item{tf}{if you supply a list of genes here the same values as for all genes are calulated for this list}

\item{genes}{see reads.taken}

\item{cutoff}{= 0.77 a good expression dataset should not use more than 77\% of the reads in the top 5\%}
}
\value{
a list of samples which have passed the test

a complicated list with al measured values
}
\description{
this check is testing how many percent of the total reads are taken from the first

x percent of the genes this function calls reads.taken.NGSexpressionSet internally.
}
