% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGSexpressionSet.R
\name{reads.taken.NGSexpressionSet}
\alias{reads.taken.NGSexpressionSet}
\title{calculates how many reads are consumed by the top genes}
\usage{
\method{reads.taken}{NGSexpressionSet}(x, percentile = 0.05, tf = NULL)
}
\arguments{
\item{x}{the NGSexpressionSet}

\item{percentile}{how many of the top genes to use (0.05)}

\item{tf}{further subselect the top genes using these genes}
}
\value{
a list of values - please read the source code of this function
}
\description{
calculates how many reads are consumed by the top genes
}

