% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_class.R
\docType{data}
\name{red}
\alias{red}
\title{reduced PMID25158935 dataset to a 100x15 StefansExpressionSet}
\format{StefansExpressionSet}
\usage{
red
}
\description{
Reduced StefansExpressionSet from the PMID25158935exp + PMID25158935samples dataset
}
\keyword{datasets}

