% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_reduce.Obj.R
\docType{methods}
\name{reduce.Obj}
\alias{reduce.Obj}
\alias{reduce.Obj,StefansExpressionSet-method}
\title{description of function reduce.Obj}
\usage{
reduce.Obj(x, probeSets = c(), name = "reducedSet")
}
\arguments{
\item{x}{the StefansExpressionSet object}

\item{probeSets}{a list of probesets to reduce the data to}

\item{name}{the new StefansExpressionSet name}
}
\description{
reduces the dataset based on genes e.g. dropps genes from the StefansExpressionSet
}

