% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpressionSet.R
\name{reduce.Obj}
\alias{reduce.Obj}
\title{reduces the dataset based on genes e.g. dropps genes from the ExpressionSet}
\usage{
reduce.Obj(x, probeSets = c(), name = "reducedSet")
}
\arguments{
\item{x}{the ExpressionSet object}

\item{probeSets}{a list of probesets to reduce the data to}

\item{name}{the new ExpressionSet name}
}
\description{
reduces the dataset based on genes e.g. dropps genes from the ExpressionSet
}

