% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeBatch.R
\docType{methods}
\name{removeBatch}
\alias{removeBatch}
\alias{removeBatch,NGSexpressionSet-method}
\title{description of function removeBatch}
\usage{
removeBatch(x, phenotype)
}
\arguments{
\item{dataObj}{the NGSexpressionSet}

\item{what}{merge by 'median','mean','sd' or 'sum'}
}
\value{
the collapsed NGSexpressionSet
}
\description{
merge the groups in the dataset into one single column The sample description is

restricted to the first entry in each group. Most variables in the sample description

table might be useless after this collape
}

