% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_reorder.obj.R
\docType{methods}
\name{reorder.genes}
\alias{reorder.genes}
\alias{reorder.genes,StefansExpressionSet-method}
\title{description of function remove.genes}
\usage{
reorder.genes(dataObj, column)
}
\arguments{
\item{dataObj}{the StefansExpressionSet object}

\item{column}{the samples column to reorder on}
}
\description{
this function reorderes the StefansExpressionSet object based on a column in the samples table (e.g. for plotting)
}

