% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_saveObj.R
\docType{methods}
\name{saveObj}
\alias{saveObj}
\alias{saveObj,StefansExpressionSet-method}
\title{description of function analyse.data}
\usage{
saveObj(data, file = NULL)
}
\arguments{
\item{file}{theoutfile default='SCExV_Grps.txt'}

\item{obj}{the Rscexv object}
}
\description{
This function saves the object either as analysis.RData or norm_data.RData if the analysi.RData has not been produced before
}

