% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_sd.filter.R
\docType{methods}
\name{sd.filter}
\alias{sd.filter}
\alias{sd.filter,stat_sd.filter-method}
\title{description of function sd.filter}
\usage{
sd.filter(x)
}
\arguments{
\item{x}{TEXT MISSING}
}
\description{
drops all genes that have an sd of 0 over the whole dataset
}

