% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_set.outpath.R
\docType{methods}
\name{set.outpath}
\alias{set.outpath}
\alias{set.outpath,StefansExpressionSet-method}
\title{description of function set.outpath}
\usage{
set.outpath(x, path = NULL)
}
\arguments{
\item{x}{the StefansExpressionSet object}

\item{path}{the new outpath (defaults to working path)}
}
\description{
Drop the samples, that have been selected!
}

