% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpressionSet.R
\docType{methods}
\name{show,ExpressionSet-method}
\alias{show,ExpressionSet-method}
\title{print the ExpressionSet}
\usage{
\S4method{show}{ExpressionSet}(object)
}
\arguments{
\item{x}{the ExpressionSet object}
}
\value{
nothing
}
\description{
print the ExpressionSet
}
\examples{

}

