% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StefansExpressionSet.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,StefansExpressionSet-method}
\title{description of function show}
\usage{
\S4method{show}{StefansExpressionSet}(object)
}
\arguments{
\item{x}{the StefansExpressionSet object}
}
\value{
nothing
}
\description{
print the StefansExpressionSet
}

