% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StefansExpressionSet.R
\docType{methods}
\name{simpleAnova}
\alias{simpleAnova}
\alias{simpleAnova,StefansExpressionSet-method}
\title{description of function simpleAnova}
\usage{
simpleAnova(x, groupCol = "GroupName", padjMethod = "BH")
}
\arguments{
\item{x}{the StefansExpressionSet object}

\item{groupCol}{the samples table column that contains the grouping information}

\item{padjMethod}{the p value correction method as described in  \code{\link[stats]{p.adjust}}}
}
\description{
This function calculates an annova to identify significant changes in the StefansExpressionSet

has a higher sensitivity for multi group analyses to identify group specific changes

or general trends in the dataset. This function adds the results into the stats slot

of the StefansExpressionSet object.
}

