% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/011_class_func.R
\docType{methods}
\name{SingleCellsNGS}
\alias{SingleCellsNGS}
\alias{SingleCellsNGS,SingleCellsNGS-method}
\title{description of function SingleCellsNGS}
\usage{
SingleCellsNGS(dat, Samples, name = "WorkingSet", namecol = "GroupName",
  namerow = "GeneID", outpath = "")
}
\arguments{
\item{dat}{data frame or matrix containing all expression data}

\item{Samples}{A sample description table}

\item{name}{The name of this object is going to be used in the output of all plots and tables - make it specific}

\item{namecol}{The samples table column that contains the (to be set) names of the samples in the data matrix}

\item{namerow}{This is the name of the gene level annotation column in the dat file to use as ID}

\item{outpath}{Where to store the output from the analysis}
}
\value{
A new SingleCellsNGS object
}
\description{
create a new SingleCellsNGS object This object is mainly used for plotting the

data
}

