% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpressionSet.R
\name{subset.ExpressionSet}
\alias{subset.ExpressionSet}
\title{subset the ExpressionSet based on a variable in the samples table}
\usage{
\method{subset}{ExpressionSet}(x, column = "Analysis", value = NULL,
  name = "newSet", mode = "equals")
}
\arguments{
\item{x}{the ExpressionSet object}

\item{column}{which column to analyze in the samples table}

\item{value}{which value to take as 'cut off'}

\item{name}{the name of the new ExpressionSet}

\item{mode}{one of 'less', 'more', 'onlyless', 'equals'}
}
\description{
subset the ExpressionSet based on a variable in the samples table
}

