% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_transpose.R
\docType{methods}
\name{t}
\alias{t}
\alias{t,02_transpose-method}
\title{description of function t}
\usage{
\S4method{t}{StefansExpressionSet}(x)
}
\arguments{
\item{x}{the StefansExpressionSet object}
}
\description{
transpose the object DELETING all stats and usedObj!
}

