% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.measurements.R
\docType{methods}
\name{update.measurements}
\alias{update.measurements}
\alias{update.measurements,SingleCellsNGS-method}
\title{description of function update.measurements}
\usage{
update.measurements(x)
}
\arguments{
\item{x}{the SingleCellsNGS object}
}
\description{
This method is inspired by PMID:24531970; it calculates the expression variance over all samples

and the fraction of cells expressing for each gene. This data is used by interesting_genes() to identify

putatively interesting genes in a SingleCellNGS object.
}
