% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.data.R
\docType{methods}
\name{write.data}
\alias{write.data}
\alias{write.data,StefansExpressionSet-method}
\title{description of function write.data}
\usage{
write.data(x, annotation = NULL)
}
\arguments{
\item{x}{the StefansExpressionSet object}

\item{annotation}{a vector of annotation data column names to include in the written table (default=none)}
}
\description{
write the StefansExpressionSet data table to disk
}
