% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpressionSet.R
\name{write.data}
\alias{write.data}
\title{write the ExpressionSet data table to disk}
\usage{
write.data(x, annotation = NULL)
}
\arguments{
\item{x}{the ExpressionSet object}

\item{annotation}{a vector of annotation data column names to include in the written table (default=none)}
}
\description{
write the ExpressionSet data table to disk
}

