% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeStatTables.R
\docType{methods}
\name{writeStatTables}
\alias{writeStatTables}
\alias{writeStatTables,StefansExpressionSet-method}
\title{description of function writeStatTables}
\usage{
writeStatTables(x, annotation = NULL, wData = F)
}
\arguments{
\item{x}{the StefansExpressionSet}
}
\description{
export the statistic files
}

