% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpressionSet.R
\name{writeStatTables}
\alias{writeStatTables}
\title{export the statistic files}
\usage{
writeStatTables(x, annotation = NULL, wData = F)
}
\arguments{
\item{x}{the NGSexpressionSet}
}
\description{
export the statistic files
}

