% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.z.score.R
\docType{methods}
\name{z.score}
\alias{z.score}
\alias{z.score,StefansExpressionSet-method}
\title{description of function z.score}
\usage{
z.score(m)
}
\arguments{
\item{m}{the matrix of column = samples and rows = genes or an StefansExpressionSet}
}
\value{
the z scored matrix
}
\description{
z score the matrix
}

