% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGSexpressionSet.R
\name{z.score.matrix}
\alias{z.score.matrix}
\title{z score the matrix}
\usage{
\method{z.score}{matrix}(m)
}
\arguments{
\item{m}{the matrix of column = samples and rows = genes}
}
\value{
the z scored matrix
}
\description{
z score the matrix
}

