% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGSexpressionSet.R
\name{z.score.NGSexpressionSet}
\alias{z.score.NGSexpressionSet}
\title{zscore the ngs dataset}
\usage{
\method{z.score}{NGSexpressionSet}(m)
}
\arguments{
\item{m}{the NGSexpressionSet}
}
\value{
The z scored NGSexpressionSet
}
\description{
zscore the ngs dataset
}

