% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.gg.melt.StefansExpressionSet.R
\docType{methods}
\name{melt}
\alias{melt}
\alias{melt,StefansExpressionSet-method}
\title{description of function melt}
\usage{
melt.StefansExpressionSet(dat, groupcol = "GroupName", colCol = NULL,
  rowCol = NULL, probeNames = NULL, na.rm = FALSE, value.name = "value")
}
\arguments{
\item{groupcol}{the column in the samples table to group the expression on}

\item{colCol}{samples column names(s) to color the grouping data}

\item{rowCol}{annotation column name(s) describing the gene groups}

\item{probeNames}{which probenames to use (ProbeSetID or Gene.Symbol ...)}

\item{data}{the StefansExpressionSet object}
}
\description{
met an StefansExpressionSet to be plotted using ggplot2 functions
}

