% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.normalize.R
\docType{methods}
\name{normalize}
\alias{normalize}
\alias{normalize,StefansExpressionSet-method}
\alias{normalize}
\alias{normalize,NGSexpressionSet-method}
\alias{normalize}
\alias{normalize,SingleCellsNGS-method}
\alias{normalize}
\alias{normalize,StefansExpressionSet-method}
\title{description of function normalize}
\usage{
normalize(object, ..., name = NULL)

\S4method{normalize}{NGSexpressionSet}(object, ..., readCounts = NULL,
  to_gene_length = FALSE, geneLengthCol = "transcriptLength", name = NULL)

\S4method{normalize}{SingleCellsNGS}(object, ..., reads = 600,
  name = "normalized")

\S4method{normalize}{StefansExpressionSet}(object, ..., name = NULL)
}
\arguments{
\item{name}{the new name of the object (deafule old name + normalized)}

\item{readCounts}{The number of reads from each bam file or another value you want to normalize the data to}

\item{to_gene_length}{FALSE whether or not to normalize the data to gene length}

\item{geneLengthCol}{the column in the annotation data.frame to (in addition) normalize the genes to (e.g. trancript length)}

\item{reads}{the required read depth}

\item{x}{The NGSexpressionSet}

\item{x}{The NGSexpressionSet}

\item{name}{the new name of the object (deafule old name + normalized)}

\item{x}{The SingleCellsNGS object}

\item{name}{the name of the new object}

\item{x}{the StefansExpressionSet object}
}
\value{
the normalized data set (original data stored in NGS$raw

the normalized data set (original data stored in NGS$raw

the normalized data set (original data stored in slot 'raw'
}
\description{
normalize the expression data (sample wise)
This generic method is implemented in a object specific way - more options available
for each obejct specific function.

normalize the expression data (sample wise)

normalize the expression data by subsampling as described in PMID 24531970

constructor that has to be implemented for a generic StefansExpressionSet
This generic version was meant for array data and I have not had the need nor time to implement this part.
}
