\name{PMID25158935exp}
\docType{data}
\title{Expression data from PMID 25158935}
\description{This data has been created using the publicly available HCS dataset from the publication 'Identification of regulatory networks in HSCs and their immediate progeny via integrated proteome, transcriptome, and DNA methylome analysis.' Cabezas-Wallscheid Cell Stem Cell 2014.
The data has been mapped against mm10 using HISAT version 0.1.5-beta 64 bit using standard paired end mapping.
The data has been quantified using Rsubread and the gtf file mm10_genes.gtf.}
\usage{PMID25158935exp}
\format{data.frame containing  24062 rows and 17 columns}
\source{http://www.ncbi.nlm.nih.gov/bioproject/PRJEB5296}
\references{
Cabezas-Wallscheid N (2014) \emph{Identification of regulatory networks in HSCs and their immediate progeny via integrated proteome, transcriptome, and DNA methylome analysis.}
Cell Stem Cell
}



