% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StefansExpressionSet.R
\docType{data}
\name{PMID25158935exp}
\alias{PMID25158935exp}
\title{Read counts for the expression data described in PMID25158935}
\format{data.frame}
\usage{
PMID25158935exp
}
\description{
The data was re-mapped against mouse mm10 using HISAT

and quantified using the R subreads package.
}
\keyword{datasets}

