% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StefansExpressionSet.R
\docType{data}
\name{PMID25158935samples}
\alias{PMID25158935samples}
\title{Read counts for the sample data for the expression information described in PMID25158935}
\format{data.frame}
\usage{
PMID25158935samples
}
\description{
The data was collected from the NCBI SRA archive
}
\keyword{datasets}

