% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.simpleAnova.R
\docType{methods}
\name{simpleAnova}
\alias{simpleAnova}
\alias{simpleAnova,StefansExpressionSet-method}
\alias{simpleAnova}
\alias{simpleAnova,SingleCellsNGS-method}
\title{description of function simpleAnova}
\usage{
simpleAnova(x, groupCol = "GroupName", padjMethod = "BH")

\S4method{simpleAnova}{SingleCellsNGS}(x, groupCol = "GroupName",
  padjMethod = "BH")
}
\arguments{
\item{x}{the StefansExpressionSet object}

\item{groupCol}{the samples table column that contains the grouping information}

\item{padjMethod}{the p value correction method as described in  \code{\link[stats]{p.adjust}}}

\item{x}{the SingleCellsNGS object}

\item{groupCol}{the samples table column that contains the grouping information}

\item{padjMethod}{the p value correction method as described in  \code{\link[stats]{p.adjust}}}
}
\description{
This function calculates an annova to identify significant changes in the StefansExpressionSet

has a higher sensitivity for multi group analyses to identify group specific changes

or general trends in the dataset. This function adds the results into the stats slot

of the StefansExpressionSet object.

This function calculates an annova to identify significant changes in the StefansExpressionSet

has a higher sensitivity for multi group analyses to identify group specific changes

or general trends in the dataset. This function adds the results into the stats slot

of the SingleCellsNGS object. In contrast to the StefansExpressionSet version of the function,

the cells showing no expression are axcluded from the stats.
}
